refactor: moved ``get_group_matches`` to general helpers
