+++ lightwood/encoder/time_series/rnn.py
@@ -17,7 +17,7 @@ from lightwood.helpers.torch import LightwoodAutocast
 from lightwood.encoder.datetime import DatetimeNormalizerEncoder
 from lightwood.encoder.time_series.helpers.rnn_helpers import EncoderRNNNumerical, DecoderRNNNumerical
 from lightwood.encoder.helpers import MinMaxNormalizer, CatNormalizer
-from lightwood.encoder.time_series.helpers.common import get_group_matches
+from lightwood.helpers.general import get_group_matches
 from lightwood.encoder.time_series.helpers.transformer_helpers import TransformerEncoder, get_chunk, len_to_mask
