+++ lightwood/helpers/general.py
@@ -4,7 +4,34 @@ import numpy as np
 import pandas as pd
 from sklearn.metrics import r2_score, f1_score, mean_absolute_error
 from lightwood.helpers.numeric import is_nan_numeric
-from lightwood.encoder.time_series.helpers.common import get_group_matches
+
+
+def get_group_matches(data, combination):
+    """Given a grouped-by combination, return rows of the data that match belong to it. Params:
+    data: dict with data to filter and group-by columns info.
+    combination: tuple with values to filter by
+    return: indexes for rows to normalize, data to normalize
+    """
+    keys = data['group_info'].keys()  # which column does each combination value belong to
+
+    if isinstance(data['data'], pd.Series):
+        data['data'] = np.vstack(data['data'])
+    if isinstance(data['data'], np.ndarray) and len(data['data'].shape) < 2:
+        data['data'] = np.expand_dims(data['data'], axis=1)
+
+    if combination == '__default':
+        idxs = range(len(data['data']))
+        return [idxs, np.array(data['data'])[idxs, :]]  # return all data
+    else:
+        all_sets = []
+        for val, key in zip(combination, keys):
+            all_sets.append(set([i for i, elt in enumerate(data['group_info'][key]) if elt == val]))
+        if all_sets:
+            idxs = list(set.intersection(*all_sets))
+            return idxs, np.array(data['data'])[idxs, :]
+
+        else:
+            return [], np.array([])
 
 
 # ------------------------- #
